PVW-2441: Animate ConfirmButtons
* Fix ConfirmButton scalability

Instead of forcing the height to 48px, adjust the animation so
that it works with any button height. Also add the 8dp vertical
padding as a default to the ConfirmButton to align with other
buttons and look consistent with (very) large text scaling
enabled.

* Use StepperIndicator for Hero in SetupSecurity

* Update ConfirmButtons parameters

Instead of passing in all the individual fields, it now
instead relies on the ConfirmButton directly, which can
be build using the new ConfirmButton.accept /
ConfirmButton.reject constructor of configured from scratch.

* Clean Up ConfirmButtons

* Add ConfirmButton.fitsOnSingleLine docs

* Clean up VerticalConfirmButtons animation

Make sure begin/end of animations align between the
horizontal and vertical variants of the ConfirmButtons.

* Fix issue in horizontal animation

When using the animation inside the SliverFillRemaining
the UI broke because the LayoutBuilder conflicted with
the SliverFillRemaining. This commit migrates it so that
it relies on fractional sizes, that way it still works
when nesting it inside a SliverFillRemaining widget.

* Animate vertical confirm buttons

* Animate horizontal confirm buttons

* Refactor ConfirmButtons for easier animation

Split up the ConfirmButtons widget so that it will be
easier to start implementing the animations.