PVW-2515: UI fixes
* Update dashboard title fontSize to 16

* Update dashboard title theme

* Update AppBarTheme

Update the AppBarTheme so that all values that can be
specified in the theme are specified there, instead of in
both the Sliver & normal WalletAppBar separately.

* Fix digid button padding issue

* Update HistoryDetailScreen AppBar

Migrate to slivers and use the collapsing SliverWalletAppbar

* Update OrganizationApprovePage

- Add divider
- Remove force vertical flag

* Update menu button font

* Always show divider on TerminalPage

The ui refresh seems to always include this divider above the buttons,
so adding it consistently.

* Update dark mode digid pin keyboard color

Make sure we use the context's colorScheme to make sure
the pin keyboard for digid also uses the correct contrasting
color in dark mode.

* Move divider on check pid page

* Update SetupSecurityCompletedPage

Update the TerminalPage so that is supports showing a divider
and use this for the SetupSecurityCompletedPage.