PVW-2236: Add support for login with BSN flow
* Rewrite is_login_flow to use Itertools::exactly_one

* Rewrite is_login_flow

* Update DisclosureBlocTest to verify login flow

* Fix OrganizationApprovePage confirm button align

Make sure buttons are stacked side by side and at the bottom of the
screen when there is enough space available.

* Update DisclosureBloc docs

* Update generated code and pubspec

* Fix formatting

* Add support for named navigation to LoginDetail

* Fix DisclosureConfirmDataAttributesPage test

Provide missing mapper

* Enable 'on report issue' button in login flow

* Localise LoginDetailScreen

* Setup LoginDetailScreen UI

* Pass on sharedDataWithOrganizationBefore flag

* Introduce LoginDetailScreen

Introduce the LoginDetailScreen and add a DisclosureCheckOrganizationForLogin
state for easier distinction between the login and the normal disclosure flow.

* Move logic to check for login flow

Move the logic to check if the current request is
a login flow into a dedicated function.

* Update copy of DisclosureSuccessPage

* Add dynamic DisclosureConfirmPinPage title

* Update OrganizationApprovePage for login

Update the ui of the OrganizationApprovePage to support
the new login flow copy.

* Update DisclosureCheckOrganization state

Update the DisclosureCheckOrganization state to include the
login flag, so that the UI can render the correct page.

* Update mock disclosure to support login type

* Provide DisclosureType through flutter_api

Provide an extra DisclosureType field through the api
to let the wallet_app know what kind of disclosure we
are dealing with (i.e. if it's a 'login' type disclosure)