PVW-2511: Update Pin Errors Accessibility
* Introduce kDefaultAnnouncementDelay

Move the announcement delay to a static const.

* Fix test

* Update Pin Errors Accessibility

- Use the Scrollable flag of the AlertDialog instead of providing our
own scroll container
- Set a dedicated semanticLabel on Android so that 'alert' is not read
out loud whenever the dialog is opened. On iOS this isn't needed (and
would actually cause the title to ALWAYS be read out loud twice, since
on ios the focus is always on the title of the dialog).
- Add a 500ms delay before all announcements so they function consistently
on iOS.