PVW-2573: Feature Check Feedback Fixes
* Tweak alignments

Update the alignments and size of the step indicator of
the app bars for more consistency.

* Add title anim to SetupSecurity success state

* Rename to ReturnUrlCase

* Update copy

* Update disclosure stop sheet copy

* Dynamic StoppedPage Copy

Show different copy in the stopped page based on the
type of login flow (standard disclosure / login). Also
introduce the LoginReturnUrlCase enum to make the code
for all the possible cases a bit more readable.

* Show PlaceholderScreen from CheckAttrs.

Instead of having the 'report issue flow', show the placeholder
screen when tapping the 'are your details incorrect' button from
inside the login flow.

* Remove pink highlight from org. button

Wrapping the InkWell in a Material Widget caused a very slight pink glaze /
highlight. Removing the Material (since its no longer needed) fixes this
UI glitch.