Add to iOS privacy manifest for user defaults
Add NSPrivacyAccessedAPICategoryUserDefaults API category with reading
and writing within app group reason.

This reason should be picked up automatically by App Store Connect, as
the permissions handler plugin includes a privacy manifest in the
bundle, which Apple currently does not. If it ever starts working, the
reason should be removed.