Add to iOS privacy manifest for disk space
Add NSPrivacyAccessedAPICategoryDiskSpace API category with writing or
deleting files reason.

The only reason this needs to be specified is that SQLCipher uses the
`statfs()` and `fstatfs()` functions for determining the filesystem
type. Of all the reasons available, this seemed to be the least
inaccurate.

See: https://github.com/sqlcipher/sqlcipher/issues/501