Verify handleError 'await' behaviour
The disclosure_bloc previously didn't await a call to handleError,
which could cause the app to crash (due to an Emitter being used after
the method that owns it completed, which is invalid behavior). This was
fixed in a previous commit, but this test makes sure that that crash is
now also caught in the the disclosure_bloc_test, making sure we will not
regress into the same error.