PVW-2329: Read out BSN digits individually
* Read out BSN digits individually

Update the screen reader so that the BSN isn't
read out as one big number, but instead read out
all the individual characters. To achieve this:

- A method was introduced to check if a string is (most
likely) a BSN. This helps us avoid pronouncing e.g. a
house number as individual digits).
- A method was introduced to take a string of numbers and
return a comma separated representation of this, to force
the screen reader to read out the numbers individually.
- The mock BSN was updated to be of a valid BSN format, so
that it passes the test of point 1.