Demo fixes
* PR Feedback

- Fix incorrect english translation
- Fix irrelevant comment
- Fix bsn tests

* Fix cards not showing up after setup

There is a situation in which the freshly added cards don't show up on the
success page after the initial setup, this happens (on non-mock builds) when you:
1. Start clean
2. Setup Wallet
3. Kill app
4. Continue Setup
5. Cards don't show up on success page

The 'fix' here simply reverts the code to what it was previously, this code was changed
because the cards are now persisted which should fix the underlying issue. But apparently
it needs some more investigation, as it only fixes the issue in the flow where the app is
not killed during the initial setup.

* Fix dutch BSN voiceover

The commas were pronounced by talkback, so using space separator
instead.

* Update translations