PVW-2719: Back behaviour during pin setup
* Simplify SetupSecurityScreen

Since we now observe the state in the build method, we can pass
that state to nested builder methods and remove the BlocBuilder
widgets. While potentially slightly less efficient performance
didn't seem to be impacted in my emulator testing, so preferring
the simplicity in code.

* Add back button to SetupPin screen

Allow the user to navigate back to the policy
screens when setting up the pin for the first time.