PVW-2377: Details Incorrect screen
* Introduce & Show InfoScreen

Show the info screen with the 'details incorrect' copy
when the user taps the button to inform the app that
the details shown on the screen are incorrect.

* Fix DataIncorrectScreen layout

Layout didn't use the bottom back button and missed a
safe area widget to make sure the content doesn't overlap
with the screen insets.