PVW-2595: Design System Buttons
* Fix test naming typo

* Test ConfirmButton states

Also removed the unused golden references for the
confirm button widget.

* Ignore warning for nested color class

* Clean up DestructiveButton

Extract the logic of darkening the color

* Fix mock build locking bug

When the user:
- Opens disclosure flow
- Triggers a timeout in the flow
- Waits for the timeout to pass
- The user ends on the dashboard (!)

The wallet should be locked in this process, code is
adjusted so it locks and the user now instead lands on the
unlock wallet page.

* Revert disclosure changes

Accidentally committed some random spaces, reverting.

* Add BottomBackButton test

* Add DestructiveButton test

* Add ListButton test

* Add LinkButton test

* Add TertiaryButton test

* Add SecondaryButton test

* Add PrimaryButton test

* Refactor LinkButton

- Make it use ButtonContent like other Buttons
- Simplify the implementation by relying only
on a custom button theme

* Remove BottomButton widget

Since the BottomButton widget mostly resembled the ListButton,
this commit opts to get rid of it completely, and instead replaces
it with usages of the ListButton widget.

* Tweak button alignment

Tweak the alignment of the buttons in the sign/issue flow.

* Exclude secondaryButton from semantics

Make sure the secondary button is excluded from the
semantics tree when it is visually hidden.

* Remove ConfirmButton

All usages have been removed, so the widget itself can now be
deleted.

* Support custom alignment in buttons

* Fix icon alignment in ConfirmActionSheet

When iconData is null we should provide a null widget
as the icon so the alignment is correct.

* Refactor LinkTile- to ListButton

The design system refers to (what we called) the LinkTileButton
as the ListButton. As such refactoring it to ListButton to align
with the design system. Furthermore the button is updated to rely
on the built in TextButton, as it's style closely resembles the
theme we use for that button. The main adjustments are the divider
support and the different shape/padding.

* Re-enable semanticsLabel

* Use Text Widget child for buttons

Change the String 'text' field inside the buttons to
be of type Text, for better compose-ability.

* Remove ConfirmButton.reject usages

Replace with the dedicated SecondaryButton widgets.

* Remove ConfirmButton.accept usages

Replace with the dedicated PrimaryButton calls.

* Remove usage of ConfirmButton()

Replace all ConfirmButton usages with dedicated instantiations
of primary/secondary/tertiary button widgets.

* Update default Buttons to support FitsWidthWidget

Instead of having to rely on the [ConfirmButton] child when rendering
the ConfirmButtons, we now implement [FitsWidthWidget] for all default
buttons so they can be used in conjunction with the ConfirmButtons
widget.

* Introduce FitsSingleLineWidget interface

Introduce an interface for children of the ConfirmButtons widget
so that the ConfirmButtons widget can figure out whether or not
to place its children horizontally or vertically.

* Adjust padding of confirm_buttons

* Introduce DestructiveButton

* Align ConfirmButton naming

* Add scenarios to ThemeScreen

Allow the developer to start all the different
flows (issue/sign/disclose/login) from the design
system page, only available for mock builds.

* Update ButtonContent width

Make sure that the buttons always take up all the
available width.

* Remove TextIconButton

TextIconButton could be replaced by the TertiaryButton,
so removing it to clean up the codebase and align with
Figma.

* Add iconPosition support to ButtonContent

* Simplify Button Setup

* Move padding to buttonTheme

* Rename Secondary -> TertiaryButton

* Add all Figma colors