PVW-2740 Add DisclosureType to WalletEvent & Remove TimelineAttribute
* Rename disclose_event.dart

* Fix ActivitySummary empty state bug

* Implement LoggedIn line for ActivitySummary

* Update HistoryDetail localizations

* Use relyingPartyOrIssuer extension

* Rename dataAttributes -> attributes

* Rename History -> WalletEventRepository

* Remove last (non l10n) timeline reference

* Rename HistoryDetailWalletEventRow

Rename the HistoryDetailWalletEventRow into WalletEventStatusHeader
to more accurately represent what it actually renders to the screen.

* Lokalise sign status text

And clean up WalletEventStatusTitleMapper

* Lokalise login status text

And clean up some more timeline references

* Remove more TimelineAttribute references

* Improve CoreHistoryRepository recent retrieval

Instead of sorting the history, we now rely on the
ordering supplied by the api to clean up the logic
to resolve the most recent disclosure/issuance events.

* Remove renamed goldens

Remove goldens that still referred to timeline so
they will eventually be re-generated under the correct
name. Also update the onPressed methods to reference
WalletEvents.

* Rename HistoryDetailTimelineAttributeRow

* Rename Section related widgets

* Remove TimelineAttribute

Delete the TimelineAttribute and its subclasses from the
codebase completely.

* Remove TimelineAttributeRepository

Update the CardDetailScreen and other dependencies of the
TimelineAttributeRepository so that it can be removed in
favor of using the WalletEvents (HistoryRepository) directly.

* Implement Equatable for WalletEvent

Make sure items can be checked for equality, this fixes
a false negative test in dashboard_bloc_test.

* Migrate ActivitySummary to WalletEvent

Update the ActivitySummary widget to work with
WalletEvents instead of TimelineAttributes, also
update the related tests

* Update Tests to use WalletEvent

Tests were still mostly referring to the TimelineAttribute classes,
they have now been updated to use the WalletEvent model, the names
still have to be updated and replaced.

* Add EventStatus to WalletEvent

Add a generic status enum to the WalletEvent and implement the
mappers that rely on this. Also update the (mock) logging logic so
a purpose is provided for issue/sign events in the log.

* Update HistoryDetailScreen to support WalletEvent

* Add WalletEvent title and description mapper

* Introduce WalletEventMappers

* Re-enable barebones TimelineAttributeRow

* Provide WalletEvent to HistoryScreen

Update the (Card)HistoryScreen to use the new WalletEvent model. Note that this is an intermediate commit that actually breaks the UI in favor of smaller commits to keep things compile-able.

* Fix provider issue

* Introduce WalletEvent usecases

* Remove Duplicate DisclosureType

* Update WalletEventMapper

Include the type mapper

* Add new WalletEventMapper

* Introduce app WalletEvent

* Add DisclosureType to WalletEvent

Also update the mock implementation to pass on the correct value. The core logic simply defaults to passing in Regular for now.