PVW-2554: Device unsupported error
* Tweak SetupSecurity Tests

Fix equality check of the error states and validate
it works as expected.

* Verify SetupSecurityBloc exposes generic errors

Make sure the SetupSecurityGenericError is exposed by
the bloc when any other unknown error is thrown by the
wallet_core.

* Verify SetupSecurityBloc exposes missing hardware

Make sure the SetupSecurityDeviceIncompatibleError is exposed by
the bloc when the corresponding error is thrown by the wallet_core.

* Verify SetupSecurityBloc exposes network errors

* Add pin verification state change tests

* Add initial pin selection tests

Verify flow for valid and invalid pins result in the expected
states.

* Add initial SetupSecurityBloc test

* Verify DeviceIncompatible warning is shown

Make sure that when the SetupSecurityScreen receives a
SetupSecurityDeviceIncompatibleError, the corresponding
error message / screen is shown.

* Navigate to Splash from Device Unsupported page

* Rename MissingKeyHardware->HardwareKeyUnsupported

* Update errorScreenDeviceIncompatibleDescription

* Introduce buildShowDetailsButton

Add a buildShowDetailsButton() method to the ErrorButtonBuilder
class so that we have a single place to construct the show details
button.

* Lokalise DeviceIncompatible Errors

* Render device not supported error

* Add core support for MissingKeyHardware error