PVW-2869: Clean up Lokalise duplicates
* Remove duplicate tests #2

Caused my incorrect 'auto resolve' of merge conflicts

* Remove duplicate tests

Caused my incorrect 'auto resolve' of merge conflicts

* Remove unused lokalise key

* Navigate to Splash from Device Unsupported page

* Rename MissingKeyHardware->HardwareKeyUnsupported

* Add unused .arb keys check to CI

Check if the arb files contain unused keys, and show a warning in
the pipeline if they do.

* Fetch latest .arbs

* Add translations_cleaner as dev dependency

And run:
- dart run translations_cleaner clean-translations

* Update errorScreenDeviceIncompatibleDescription

* Introduce buildShowDetailsButton

Add a buildShowDetailsButton() method to the ErrorButtonBuilder
class so that we have a single place to construct the show details
button.

* Lokalise DeviceIncompatible Errors

* Render device not supported error

* Add core support for MissingKeyHardware error

* Fix linebreak issue

* Add 'Show Details' tests

Make sure the correct 'show details' cta is shown
on the errorscreen and errorpage.

* Make sure the correct info button is shown

* Verify server error for DisclosureScreen

* Verify server error for WalletPersonalizeScreen

* Verify server error for SetupSecurityScreen

* Change Default ErrorCtaStyle to retry

Since we mostly want to show the try again cta, default
the ErrorCtaStyle to retry so that is shown by default.

* Update ServerError UI

- Update the UI to show the new headline/description.
- Show the retry cta when coming from the pin_screen.
- Validate correct server error state is shown from the
pin_screen when receiving the network error with
hasInternet=true.

* Simplify ErrorPage api

Since there is not a lot customization needed (yet) for
these pages, prefer keeping the api as simple as possible
by deciding what button and text to show based on the
cta style inside the ErrorPage widget.

* Update ErrorPage.network api

Update the api to work with the buttons directly instead of
a list of parameters.

* Update ErrorPage.generic api

- Update the api to work with the buttons directly instead of
a list of parameters
- Verify that DisclosureScreen shows the correct generic error

* Validate generic error for WalletPersonalizeGenericError

* Validate generic error for SetupSecurityScreen

* Validate generic error for PinScreen

* Fix typo in test name

* Add missing SafeAreas

* Fix error_screen and confirm_buttons tests

* Verify internet error for DisclosureScreen

* Verify internet error for WalletPersonalizeScreen

* Verify internet error for SetupSecurityScreen

* Clean up pin_screen_test

* Verify PinValidateNetworkError

Verify that the PinValidateNetworkError with hasInternet=false
on the PinScreen results in the correct error being shown.

Thies is, no internet title and a try again button.

* Add initial pin_screen_test

* Tune ErrorPage layout

Update icon and spacings.

* Update ErrorPage layout

Match the layout (and api) of the refreshed ErrorScreen.

* Add ScrollBar to ErrorScreen

* Update actions on ErrorScreen

Dynamically show the close button based on the
ErrorCtaStyle.

* Update ErrorScreen showGeneric

* Update version widgets styling

Change to dedicated prefix and value text styles
for easier and more predictive customization.

* Update version widgets

Separate the prefix and value in lokalise so
we can easily use custom weight for both texts.

* Fix error_screen_test.dart

* Make no internet description dynamic

* Pass buttons to ErrorScreen

Instead of passing individual parameters update the
api to accept the buttons directly, also tweak the
showNoInternet() accordingly.

* Remove title from ErrorScreen

* Update layout of ErrorScreen

Add collapse-able layout and generic show details
cta for the internet error.

* Allow flipping of vertical_confirm_buttons

* Tweak ConfirmButtons

Vertical: Make sure init/dispose is called at the correct time

Horizontal: Make sure init/dispose is called a the correct time
and tweak the alignment so that the secondary button is offscreen
when its hidden and the buttons are bottom aligned when more space
is available.

* Tweak OtherStylesTab

Avoid duplicate ids and thereby eliminate the heroes
duplicate key error printed to the log.

* Update SliverWalletAppBar leading logic

Update it so that the automaticallyImplyLeading logic matches
that of the WalletAppBar.

* Add ErrorScreens to OtherStylesTab

* Clean up HelpSheet

For more consistency:
- Use ListButton
- Use BottomCloseButton

* Introduce ErrorDetailsSheet

* Update HelpSheet to include config version