PVW-2656: Store metadata
* Add basic store explanations

* Add character limit to titles

Add the character limits of each field to the title,
for easier referencing when editing them in place.

* Update app_icon

Update the app icons with the newly provided (correctly
sized) alternatives.

* Add basic AppStore texts

Add app store texts, no unique texts were preconfigured
in the app store connect console the play store texts have
been re-used. Also removed the line breaks from all descriptions
to allow the play- or appstore to flow the text as desired.

* Add basic PlayStore texts

* Resize feature_images

Supplied images were factor 3 to large, resized them
so that they are supported by the play store.

* Restructure files

Restructure how the files are organized and rename them
to consistently use lowercase naming.

* Basic Store Metadata

Commit an initial structure, together with the images
to the repository.