PVW-2831: Dynamic Gender Icon
* Remove unused l10n reference

* Remove stale comment

* Verify Pid Preview Icons

Verify the correct icons are shown for the male/female
gender types.

* Update Mock Pid Data

Make sure we supply the gender as a GenderValue from the mock
as well, as we now rely on that to render the correct icon
in the pid preview screen.

* Dynamic Gender Icon

Update the PidAttributeMapper to select the icon used in the
gender row dynamically, i.e. based on the gender attribute.