PVW-2910: Update PlaceholderScreen & api
* Add Placeholder.showHelp()

Add a new variant of the placeholder screen for the
show help / faq functionality. Also update the HelpIconButton
to show this variant of the placeholder screen when it's
pressed. Finally make sure automaticallyImplyLeading for
the PlaceHolderScreen is set to true, as the back button
should be shown when the user can navigate back.

* Add missing placeholder svg

* Update translations

* Update PlaceholderScreen & api

- Update the ui of the placeholder screen to match
the new design.
- Update the api (.showXxx) of the PlaceholderScreen
to match that of the ErrorScreen.