Improve ephemeral ID processing in Verifier
This moves the parsing of the verifier URL query parameters from
`wallet_server` to the `mdoc` crate. This prevents the called verifier
URL from being parsed and deconstructed in `wallet_server`, only to be
re-assembled again in the `Verifier`. Since `axum` knows what URL was
called, we are able to pass that directly to the `Verifier`.

This has the added advantage of the formatting and parsing of these
query parameters being done in the same place, as opposed to having this
split over two different crates.