PVW-2833: Handle blocked external scanner
* Shorten handleError method

* Add ResetWalletDialog UI Tests

Noticed the ResetWalletDialog also wasn't tested yet,
as such adding some basic tests for this widget too,
since they are quite similar to the ScanWithWalletDialog
tests and thus easy to implement.

* Add ScanWithWallet UI Tests

* Add tests for SourceMismatchError

Add tests that check if the correct state is emitted based on the
thrown CoreDisclosureSourceMismatchError.

* Fix CardDetail Scroll issue

Noticed a minor scroll issue on the CardDetailScreen,
where the list of actions can be scrolled separately from
the page content. The fix is very minor so opted to commit it
on my working branch.

* Hande Disclosure cancel while loading

If the user would cancel the disclosure flow while the state
was loading, some weird states could occur, mainly the stop
sheet appearing with a '-' where the organization name should be,
and the app crashing in debug mode because of missing reporting
options.

We now gracefully handle this case by adding dedicated copy when
the org. name is not yet available, and providing default (empty)
implementations for the ReportingOptions.

* Introduce ScanWithWalletDialog

Add the ScanWithWalletDialog which is shown when the
CoreDisclosureSourceMismatchError is thrown, indicating
that a QR code was scanned with an external app in stead
of the NL Wallet's built in scanner. It redirects the user
to the internal QR scanner (or allows the user to simply
dismiss the warning).

* Add CoreDisclosureSourceMismatchError

Add a new error type, namely the CoreDisclosureSourceMismatchError
which bubbles up when the wallet_core throws an error of type
'DisclosureSourceMismatch'.