PVW-2955: Lock dashboard crossAxisCount
* Lock dashboard crossAxisCount

Instead of letting the crossAxisCount grow infinitely based
on the width of the screen, we now max it out at _kMaxCrossAxisCount
(currently set to 2). This fixes the issue where the UI looks weird
when it would theoretically fit 3 cards, but the wallet only contains 2.
It also helps with accessibility, as cards can grow to be quite large.