PVW-2308 / PVW-2305: ScreenReader tweaks
* Conditionally hide gov. logo

Hide the government logo in the semantics tree when the user
is passed the initial introduction page.

* Fix onboarding voiceover focus issue

In the onboarding, the focus would not jump to the top of the
screen when the user was relying on voiceover to navigate through
the app. By setting a dynamic key on the next button, we force flutter
to reconsider what widget to give focus, and since a key with that button
no longer exists it restarts at the top of the screen.

* Conditionally disable FakePager animations

When the user has a screen reader enabled, disable the fake
paging animations, so not to interfere with the screen reader.

* Fix confirm attributes title tap area