PVW-3030: App bar transition Tweaks
* Remove SliverWalletAppBarCompanion

* Pass on scrollControllers

Pass on the ScrollController to all SliverWalletAppBars
so that the animation range can be based on the
maxScrollExtent.

* Introduce SliverWalletAppBarCompanion

A sliver widget that can be placed at the bottom of
a CustomScrollView to make sure the SliverWalletAppBar
has all the scrollRange it might need to complete it's
transition animation.

* Fix ScrollController nullpointer

Trying to resolve the maxScrollExtent too early could
cause a nullpointer exception. These extra validations
should help avoid that error.

* Fix 'stuck' opacity for SliverWalletAppBar

Check in with the scrollController to see if the maxScrollExtent
is sufficient to perform the complete opacity animation. If it is
not we speed up the opacity animation by taking this into account.

* Fix 'stuck' opacity for FadeInAtOffset

Instead of statically defining the appear and fully visible offset,
the fully visible offset now also takes the maximum scrollrange into
account (when available). This makes sure that even if the scroll cant
reach the fully visible offset the child will still become fully visible
when the content is fully scrolled.