PVW-3008: Personalize scroll behaviour fix
* Add ScrollOffsetProvider

On rotation the WalletPersonalizeScreen sometimes failed
to re-observe the scroll controller. Now providing the offset
from the top of the tree for consistency of the fade in title
behaviour.

* Add collapsing title to personalization

The wallet_personalize_screen was missing the collapsing header in certain
screens. This is added for the error and success pages with this commit.