PVW-2563: Update History Details
* Clean up translations

* Fix ScrollController conflict

* Lokalise 'details' CTA

* Verify HistoryDetailScreen pages

Verify that the correct page is rendered based on the supplied
wallet event type.

* Remove unused docType

DocType attributes was passed around but not used, cleaning it up
to reduce complexity.

* Remove unused code

* Add issue title localization

* Add details & report CTAs

* Add visibility logic to pages

Decide whether or not to show certain slivers into the pages, as that
logic may vary per page and be independent of what is passed to the
builder to create the widget in the first place.

Add buildAboutOrganizationSliver to show the about [org] cta.

* Add content to issue card page

* Add policy sliver

* Lokalise attributes section

Also add the attributes and purpose statement to the login view

* Display attributes on detail pages

* Display purpose on detail pages

* Add StatusHeader to Detail Pages

* Add timestamp subtitle to pages

* Break out detail pages

Instead of building all variants of the UI in the same
screen widget, delegate this to individual pages to keep
the code more readable.

* Rename file to match widget

* Update EventRow

* Update HistoryDetails header and EventRow

* Fix SharedAttributesCard header

Make sure the header background fills the complete width,
e.g. wasn't scaling correctly for driving license card.

* Make HistoryDetailScreen title dynamic