PVW-3080: Update UI of about org CTA
* Add HistoryDetailCommonBuilders tests

Make sure that a logo is rendered on the 'about org'
cta. And validate the basics of all other widget builders
inside HistoryDetailCommonBuilders.

* Update maxScrollExtent getter

Make it more resilient by not forcing the lookup
through the first attached client. Useful when in the
brief moment more than one client might be attached, e.g.
when dismissing the 'locked' overlay.

* Update about organization button

- Add support for trialing widget to ListButton, also
tweak the sizes to match the updated design.
- Render the organization logo as the trailing widget
in the 'About organization' button.