PVW-2742: Fix duplicate PID issuance (wallet_app)
* Verify issuance url error flow

* Add bloc test for pid issuance

* Verify basic WalletPersonalizeBloc states

Also fix issue with some assertions in other tests, where
the 'expect' was missing causing the tests to never fail.

* Update wallet_personalize_bloc for PVW-2742

Check if the pid is already added to the users wallet when the bloc is
created. This makes sure the correct 'cards added' state is rendered
when the user got locked out of the wallet (due to a auto lock timeout)
when looking at the 'cards successfully added' screen. See PVW-2742 for
the exact issue and steps to reproduce.

* Avoid recreating consumable

By placing the instantiation of the Consumable in the builder bloc,
it would get re-created whenever the builder was called, defeating the
purpose of the consumable. We now instantiate the consumable before the
builder is created (i.e. outside the builder) to avoid this issue.