PVW-2951: Accessibility tweaks
* Refactor method

Update method name to represent what it actually returned

* Update PolicyEntryRow semantics

- Mark a policy view as being a link when it has a onTap
action. This helps on iOS to inform the user a section is
tappable, as on iOS the onTapHint is ignored.
- Update the dutch 'generalWCAGOpenLink' so that the pronounced
text on android makes more sense, as the systems prefix is
'dubbel tik om te' which resulted in the weird 'dubbeltik om te open link'
with the old copy.

* Merge HistoryDetail Semantics

Merge title and body of widgets build by the detail
common builders according to the spec. PVW-3095.

* Introduce paragraphed_sliver_list

Add a slightly more efficient way of showing
the paragraphed list in a sliver setup. Intended
as a fix for an accessibility issue, which it
unfortunately did not fix. But kept for efficiency.

* Add option to hide text on WalletCard

On the CheckAttributesScreen the text on the miniature
wallet card should be hidden. This commit adds support
for that behaviour to the WalletCardItem widget and
enables that flag on the CheckAttributesScreen.

* Pop the MenuScreen onLock

- Introduce the LockStateMixin for easy observation
of the lock state on a stateless widget, so we can
act upon changes to this state.
- Pop the MenuScreen on lock events (PVW-3104)

* Use systemLocale for tapHint

The taphint is always announced together with a announcement from
the system language. E.g. "Double tap [msg]" or "Dubbeltik om [msg]",
this change detects the system locale and uses those localisations to
form the tapHint.

* Provide maxScrollExtent with ScrollOffset

Provide the maxScrollExtent with the ScrollOffset
so that the FadeInAtOffset animation can be sped
up when it relies on the ScrollOffset too.

* Fix duplicate title

When font scaling allowed the title to move into
the appbar on the wallet_personalize_screen in the
success state, it could show up twice because the
scrollOffset of the previous page was used to calculate
the visibility. This is now fixed by resetting the provided
offset when navigating to the next page.