PVW-2298: Open return_url on disclosure cancellation & errors
* Add missing formatter comma

* Verify returnUrl is available

Verify that returnUrl is provided in the disclosure-
stopped and error states.

* Add trailing commas for linter

* Update CoreErrorMapper tests

Also regenerate code for new server json annotation

* Add initial support for Server type error

The wallet core can also throw Server errors now, these do not yet
require special handling, so we simply treat them as GenericErrors for
now. But this change makes sure the error mapper doesn't crash when
they occur.

* Update lambda callback naming

Consistently name what is coming back through
the `handleError` callbacks `error` vs a mix of
`ex` and `error`.

* Add returnUrl to DisclosureGenericError

Add the returnUrl to the DisclosureGenericError state (when
it's available) so the app can redirect the user back to the
browser in case of 'recoverable' errors.

* Pass on returnUrl on cancellation

Pass on the returnUrl to the DisclosureScreen when the
state moves to the stopped state. This way the user can be
redirected to a 'stopped' web page, even when the disclosure
session is cancelled by the user itself.

* Introduce kotlin like let operator