PVW-3134: Improve dart test code coverage
* Add missing consts

* Update WalletPersonalizeBloc tests

* Update WalletPersonalizeBloc tests

* Update WalletPersonalizeScreen tests

* Add SplashBloc test

* Improve SignScreen tests

* Add QrNoPermission test

* Add QrBloc tests

* Introduce CheckHasPermissionUseCase

Rewrite the QrBloc so that it checks the current
permission status through an exernal usecase, so that
we can mock the result when testing the behaviour of
the QrBloc.

* Add LoginDetailScreenArgument test

* Add ImageMapper tests

* Update IssuanceScreen tests

Validate that page corresponding to state is rendered.

* Add MenuBloc test

* Add MenuScreen test

* Add IssuanceBloc tests

Add basic tests for the issuance bloc to validate the
happy path.

* Add test for HistoryOverviewLoadFailure state

* Ignore mock_digid files in coverage report

* Update IntroductionScreen tests

* Update IntroductionPrivacyScreen tests

* Update IntroductionConditionsScreen tests

* Add InfoScreen test

* Add test for launch_util

* Update wallet_mocks

Make sure the service locator contains all the generated nice
mocks, so they can all be fetched through Mocks.create().

* PVW-3130: Increase dart unit test coverage #2

* Clean unused translations

* Add tests for CheckPermissionOnResume

* Update DisclosureScreen tests

Also tweak the walletAppWrapper to take generic
providers so provided dependencies can be scopes
above the navigator (i.e. so that they persist
through navigation to other routes).

* Update HelpSheet test

* Improve withDependency

Instead of defining an explicit builder, use the built in
builder parameter of the Provider widget.

* Add basic SignBloc test

* Update DisclosureScreenArgument test

* Fix DigidHelpScreen test

* Add Environment test

* Update DigidHelpScreen test

* Update DataIncorrectScreen test

* Fix NumberedList issue

On large font scaling where the user would want to scroll
content with a numberedlist, she might not be able to scroll
as the scroll gesture is caught by the numberedlist instead
of its container. Now setting NeverScrollablePhysics to avoid
this behaviour.

* Add DashboardScreenArgument test

* Improve WalletCardItem tests

* Update ThemeScreen tests

* Update PlaceholderScreen tests

* Add ErrorDetailsSheet test

* Add HistoryOverviewBloc tests

* Update HistoryDetailScreenArgument test

* Update HistoryDetailScreen tests

* Add HistoryDetailBloc tests

* Add test for CardAttributeRow

* Clean unused translations

* Add tests for CheckPermissionOnResume

* Update DisclosureScreen tests

Also tweak the walletAppWrapper to take generic
providers so provided dependencies can be scopes
above the navigator (i.e. so that they persist
through navigation to other routes).

* Update HelpSheet test

* Improve withDependency

Instead of defining an explicit builder, use the built in
builder parameter of the Provider widget.

* Add basic SignBloc test

* Update DisclosureScreenArgument test

* Fix DigidHelpScreen test

* Add Environment test

* Update DigidHelpScreen test

* Update DataIncorrectScreen test

* Fix NumberedList issue

On large font scaling where the user would want to scroll
content with a numberedlist, she might not be able to scroll
as the scroll gesture is caught by the numberedlist instead
of its container. Now setting NeverScrollablePhysics to avoid
this behaviour.

* Add DashboardScreenArgument test

* Improve WalletCardItem tests

* Update ThemeScreen tests

* Update PlaceholderScreen tests

* Add ErrorDetailsSheet test

* Fix lint errors

* Update CardDetailScreen tests

* Add LockStateMixin tests

Actually discovered a bug in the stream (wrong
distinct operator), fixed together with these
tests.

* Add CheckAttributesBloc tests

* Add ChangeLanguageBloc tests

* Improve CardHistoryScreen tests

* Add CardHistoryBloc tests

* Improve CardDetailScreen tests

* Add CardDetailBloc tests

* Improve CardDataScreen test

* Add test for CardDataIncorrectScreen

* Improve CardDataScreenArgument test

* Improve ObserveWalletCardDetailUseCaseImpl test

* Improve coverage of PidIssuanceStatus

* Verify Document equals implementation

* Remove toString from AppImageData

No longer actively used and the props will also
supply a sane toString() implementation.

* Validate WalletEvent equals

* Validate ContinueIssuanceResult getters

* Validate StartIssuanceResult getters

* Add CardDataBloc tests